# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TeaType {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  createdAt: string;
}

export interface TeaTypeInput {
  /** @minLength 1 */
  name: string;
  category: string;
  description?: string;
  color?: string;
}

export interface TeaTypeUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  description?: string;
  color?: string;
}

export interface BrewingMethod {
  id: number;
  name: string;
  /** @nullable */
  temperatureCelsius?: number | null;
  /** @nullable */
  steepTimeSeconds?: number | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface BrewingMethodInput {
  /** @minLength 1 */
  name: string;
  temperatureCelsius?: number;
  steepTimeSeconds?: number;
  description?: string;
}

export interface BrewingMethodUpdate {
  /** @minLength 1 */
  name?: string;
  temperatureCelsius?: number;
  steepTimeSeconds?: number;
  description?: string;
}

export interface TeaSession {
  id: number;
  /** @nullable */
  teaTypeId?: number | null;
  /** @nullable */
  brewingMethodId?: number | null;
  cups: number;
  /** @nullable */
  notes?: string | null;
  loggedAt: string;
  createdAt: string;
  teaType?: TeaType;
  brewingMethod?: BrewingMethod;
}

export interface TeaSessionInput {
  teaTypeId?: number;
  brewingMethodId?: number;
  /** @minimum 1 */
  cups: number;
  notes?: string;
  loggedAt?: string;
}

export interface TeaSessionUpdate {
  teaTypeId?: number;
  brewingMethodId?: number;
  /** @minimum 1 */
  cups?: number;
  notes?: string;
  loggedAt?: string;
}

export interface StatsOverview {
  today: number;
  thisWeek: number;
  thisMonth: number;
  allTime: number;
  totalCupsToday: number;
  totalCupsWeek: number;
  totalCupsMonth: number;
  totalCupsAllTime: number;
}

export interface TeaTypeStat {
  teaTypeId: number;
  teaTypeName: string;
  category?: string;
  /** @nullable */
  color?: string | null;
  sessionCount: number;
  totalCups: number;
}

export interface DailyStat {
  date: string;
  sessionCount: number;
  totalCups: number;
}

export type ListTeaSessionsParams = {
limit?: number;
offset?: number;
};