# lib/api-zod/src/generated/types/brewingMethod.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface BrewingMethod {
  id: number;
  name: string;
  /** @nullable */
  temperatureCelsius?: number | null;
  /** @nullable */
  steepTimeSeconds?: number | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}