# lib/api-zod/src/generated/types/brewingMethodInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface BrewingMethodInput {
  /** @minLength 1 */
  name: string;
  temperatureCelsius?: number;
  steepTimeSeconds?: number;
  description?: string;
}