# lib/api-zod/src/generated/types/dailyStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyStat {
  date: string;
  sessionCount: number;
  totalCups: number;
}