# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOverview {
  today: number;
  thisWeek: number;
  thisMonth: number;
  allTime: number;
  totalCupsToday: number;
  totalCupsWeek: number;
  totalCupsMonth: number;
  totalCupsAllTime: number;
}