# lib/api-zod/src/generated/types/teaSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { BrewingMethod } from './brewingMethod';
import type { TeaType } from './teaType';

export interface TeaSession {
  id: number;
  /** @nullable */
  teaTypeId?: number | null;
  /** @nullable */
  brewingMethodId?: number | null;
  cups: number;
  /** @nullable */
  notes?: string | null;
  loggedAt: string;
  createdAt: string;
  teaType?: TeaType;
  brewingMethod?: BrewingMethod;
}