# lib/api-zod/src/generated/types/teaSessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TeaSessionInput {
  teaTypeId?: number;
  brewingMethodId?: number;
  /** @minimum 1 */
  cups: number;
  notes?: string;
  loggedAt?: string;
}