# lib/api-zod/src/generated/types/teaType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TeaType {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  createdAt: string;
}