# lib/api-zod/src/generated/types/teaTypeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TeaTypeInput {
  /** @minLength 1 */
  name: string;
  category: string;
  description?: string;
  color?: string;
}