# lib/api-zod/src/generated/types/teaTypeStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tea Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TeaTypeStat {
  teaTypeId: number;
  teaTypeName: string;
  category?: string;
  /** @nullable */
  color?: string | null;
  sessionCount: number;
  totalCups: number;
}